[wip] job results representation negotiation
